+++ js/piwik.js
@@ -994,7 +994,7 @@ if (typeof JSON_PIWIK !== 'object' && typeof window.JSON === 'object' && window.
     disableCrossDomainLinking, isCrossDomainLinkingEnabled,
     addListener, enableLinkTracking, enableJSErrorTracking, setLinkTrackingTimer, getLinkTrackingTimer,
     enableHeartBeatTimer, disableHeartBeatTimer, killFrame, redirectFile, setCountPreRendered,
-    trackGoal, trackLink, trackPageView, trackRequest, trackSiteSearch, trackEvent,
+    trackGoal, trackLink, trackPageView, getNumTrackedPageViews, trackRequest, trackSiteSearch, trackEvent,
     setEcommerceView, addEcommerceItem, trackEcommerceOrder, trackEcommerceCartUpdate,
     deleteCookie, deleteCookies, offsetTop, offsetLeft, offsetHeight, offsetWidth, nodeType, defaultView,
     innerHTML, scrollLeft, scrollTop, currentStyle, getComputedStyle, querySelectorAll, splice,
@@ -3186,7 +3186,11 @@ if (typeof window.Piwik !== 'object') {
                 // Domain hash value
                 domainHash,
 
-                configIdPageView;
+                configIdPageView,
+
+                // we measure how many pageviews have been tracked so plugins can use it to eg detect if a
+                // pageview was already tracked or not
+                numTrackedPageviews = 0;
 
             // Document title
             try {
@@ -6543,6 +6547,13 @@ if (typeof window.Piwik !== 'object') {
             };
 
             /**
+             * Get the number of page views that have been tracked so far.
+             */
+            this.getNumTrackedPageViews = function () {
+                return numTrackedPageviews;
+            };
+
+            /**
              * Log visit to this page
              *
              * @param string customTitle
@@ -6558,6 +6569,7 @@ if (typeof window.Piwik !== 'object') {
                     });
                 } else {
                     trackCallback(function () {
+                        numTrackedPageviews++;
                         logPageView(customTitle, customData, callback);
                     });
                 }